requirements.txt
faiss-cpu==1.7.2
fastapi==0.85.0
SQLAlchemy==1.4.41
aiofiles==22.1.0
asyncpg==0.26.0
uvicorn==0.18.3
pydantic==1.10.2
psycopg2-binary
python-multipart